+++ CHANGELOG.md
@@ -535,6 +535,18 @@ v2.13.0 (2026-03-30)
   - Added the `advanced_search_to_fif.ipynb`, `yara_hunt_to_fif.ipynvb` and `email_file_reputation` notebooks.
 
 
+
+v2.14.0 (2026-06-12)
+-------------------
+### Improvements
+- **ticloud** module
+  - Added an optional `private` argument to the methods of `URLThreatIntelligence` and `NetworkReputation` classes.
+
+- **advanced** module
+  - Added the `enriched_malware_detection_feed`, `enriched_mwp_change_feed`, `enriched_new_files_feed`, `enriched_network_iocs_feed` and `enriched_data_change_feed` methods to the `AdvancedActions` class.
+
+  
+
 -------------------